[FIX,API] fix blast statistics after changes for 2.3
In SeqAn 2.3 the blast module was adapted to use the
BlastRecord's qLength member instead of the BlastMatch's
qLength member (to remove redundancy).
But the statistics module was forgotten. This patch
adapts the statistics model, preserving the original
interface with the deprecated attribute and introducing
the new one.
=> no hard breakage, but warnings will suggest changing
to the new interface.